Enum alloy_sol_types::PanicKind

#[repr(u32)]
pub enum PanicKind {
    Generic,
    Assert,
    UnderOverflow,
    DivisionByZero,
    EnumConversionError,
    StorageEncodingError,
    EmptyArrayPop,
    ArrayOutOfBounds,
    ResourceError,
    InvalidInternalFunction,
}

Represents a Solidity panic. Same as the Solidity definition.

Variants

Generic

Generic / unspecified error.

Generic compiler inserted panics.

Assert

Used by the assert() builtin.

Thrown when you call assert with an argument that evaluates to false.

UnderOverflow

Arithmetic underflow or overflow.

Thrown when an arithmetic operation results in underflow or overflow outside of an unchecked { ... } block.

DivisionByZero

Division or modulo by zero.

Thrown when you divide or modulo by zero (e.g. 5 / 0 or 23 % 0).

EnumConversionError

Enum conversion error.

Thrown when you convert a value that is too big or negative into an enum type.

StorageEncodingError

Invalid encoding in storage.

Thrown when you access a storage byte array that is incorrectly encoded.

EmptyArrayPop

Empty array pop.

Thrown when you call .pop() on an empty array.

ArrayOutOfBounds

Array out of bounds access.

Thrown when you access an array, bytesN or an array slice at an out-of-bounds or negative index (i.e. x[i] where i >= x.length or i < 0).

ResourceError

Resource error (too large allocation or too large array).

Thrown when you allocate too much memory or create an array that is too large.

InvalidInternalFunction

Calling invalid internal function.

Thrown when you call a zero-initialized variable of internal function type.

Implementations

impl PanicKind

pub const fn from_number(value: u32) -> Option<Self>

Returns the panic code for the given number if it is a known one.

pub const fn as_str(self) -> &'static str

Returns the panic code’s string representation.