Macro alloy_sol_types::sol

sol!() { /* proc-macro */ }

Parses Solidity syntax to generate types that implement alloy-sol-types traits.

These types may then be used for safe ABI and EIP-712 encoding and decoding.

Examples

Note: the following examples cannot be tested here because the generated code references alloy-sol-types, so they are [tested in that crate] tests and included with include_str! in this doc instead.

Structs and enums

use alloy_primitives::{Address, U256};
use alloy_sol_types::{sol, SolType};
use hex_literal::hex;

// Struct definitions will generate a struct with the same name and fields.
// No casing convention is enforced.
sol! {
    struct Foo {
        uint256 bar;
        address[] baz;
    }

    /// Nested struct.
    struct Nested {
        Foo[2] a;
        address b;
    }

    // TODO: enums
    /*
    enum Enum {
        A,
        B,
        C,
    }
    */
}

#[test]
fn structs() {
    let my_foo = Foo {
        bar: U256::from(42),
        baz: vec![Address::repeat_byte(0x11), Address::repeat_byte(0x22)],
    };

    let _nested = Nested {
        a: [my_foo.clone(), my_foo.clone()],
        b: Address::ZERO,
    };

    let abi_encoded = Foo::encode(&my_foo);
    assert_eq!(
        abi_encoded,
        hex! {
            "000000000000000000000000000000000000000000000000000000000000002a" // bar
            "0000000000000000000000000000000000000000000000000000000000000040" // baz offset
            "0000000000000000000000000000000000000000000000000000000000000002" // baz length
            "0000000000000000000000001111111111111111111111111111111111111111" // baz[0]
            "0000000000000000000000002222222222222222222222222222222222222222" // baz[1]
        }
    )
}

UDVT and type aliases

use alloy_primitives::Address;
use alloy_sol_types::{sol, SolType};

// Type definition: generates a new struct that implements `SolType`
sol! {
    type MyType is uint256;
}

// Type aliases
type B32 = sol! { bytes32 };
// This is equivalent to the following:
// type B32 = alloy_sol_types::sol_data::Bytes<32>;

type SolArrayOf<T> = sol! { T[] };
type SolTuple = sol! { tuple(address, bytes, string) };

#[test]
fn types() {
    let _ = <sol!(bool)>::encode_single(&true);
    let _ = B32::encode_single(&[0; 32]);
    let _ = SolArrayOf::<sol!(bool)>::encode_single(&vec![true, false]);
    let _ = SolTuple::encode_single(&(Address::ZERO, vec![0; 32], "hello".to_string()));
}

Functions and errors

use alloy_primitives::{keccak256, U256};
use alloy_sol_types::{sol, SolCall, SolError};
use hex_literal::hex;

// Unnamed arguments will be given a name based on their position,
// e.g. `_0`, `_1`...
//
// A current limitation for these types is that custom types, like structs,
// must be defined in the same macro scope, otherwise a signature cannot be
// generated at compile time.

sol! {
    /// Function definitions generate two types that implement [`SolCall`]:
    /// 1. `<name>Call`: struct with the function arguments;
    /// 2. `<name>Return`: struct with the return values;
    /// where `<name>` is the case-preserved name of the function.
    ///
    /// In the case of overloaded functions, an underscore and the index of the
    /// function will be appended to `<name>` (like `foo_0`, `foo_1`...) for
    /// disambiguation, but the signature will remain the same.
    ///
    /// Both of these types will have the attributes of the function, like this
    /// doc comment, but this might change in the future.
    ///
    /// Currently, return structs should only be used for decoding data using
    /// `decode_raw`, as the generated signature is not valid.
    function foo(uint256 a, uint256 b) external view returns (uint256);

    // These will be interpreted as `overloaded_0`, `overloaded_1`, and
    // `overloaded_2`, but the signatures will be the same.
    function overloaded();
    function overloaded(uint256) returns (uint256);
    function overloaded(string);

    /// Implements [`SolError`].
    #[derive(Debug, PartialEq)]
    error MyError(uint256 a, uint256 b);
}

#[test]
fn function() {
    assert_call_signature::<fooCall>("foo(uint256,uint256)");

    // not actually a valid signature, and it shouldn't be relied upon for
    // ABI encoding
    assert_call_signature::<fooReturn>("foo(uint256)");

    let call = fooCall {
        a: U256::from(1),
        b: U256::from(2),
    };
    let _call_data = call.encode();

    // the signatures are unaffected
    let _ = overloaded_0Call {};
    assert_call_signature::<overloaded_0Call>("overloaded()");

    let _ = overloaded_1Call { _0: U256::from(1) };
    let _ = overloaded_1Return { _0: U256::from(2) };
    assert_call_signature::<overloaded_1Call>("overloaded(uint256)");

    let _ = overloaded_2Call { _0: "hello".into() };
    assert_call_signature::<overloaded_2Call>("overloaded(string)");
}

#[test]
fn error() {
    assert_error_signature::<MyError>("MyError(uint256,uint256)");
    let call_data = hex!(
        "0000000000000000000000000000000000000000000000000000000000000001"
        "0000000000000000000000000000000000000000000000000000000000000002"
    );
    assert_eq!(
        MyError::decode_raw(&call_data, true),
        Ok(MyError {
            a: U256::from(1),
            b: U256::from(2)
        })
    );
}

fn assert_call_signature<T: SolCall>(expected: &str) {
    assert_eq!(T::SIGNATURE, expected);
    assert_eq!(T::SELECTOR, keccak256(expected)[..4]);
}

fn assert_error_signature<T: SolError>(expected: &str) {
    assert_eq!(T::SIGNATURE, expected);
    assert_eq!(T::SELECTOR, keccak256(expected)[..4]);
}

Events

#![allow(clippy::assertions_on_constants)]

use alloy_primitives::{keccak256, B256, U256};
use alloy_sol_types::{sol, token::WordToken, SolEvent};
use hex_literal::hex;

sol! {
    #[derive(Default)]
    event MyEvent(bytes32 indexed a, uint256 b, string indexed c, bytes d);

    event LogNote(
        bytes4   indexed  sig,
        address  indexed  guy,
        bytes32  indexed  foo,
        bytes32  indexed  bar,
        uint              wad,
        bytes             fax
    ) anonymous;

    struct Data {
        bytes data;
    }
    event MyEvent2(Data indexed data);
}

#[test]
fn event() {
    assert_event_signature::<MyEvent>("MyEvent(bytes32,uint256,string,bytes)");
    assert!(!MyEvent::ANONYMOUS);
    let event = MyEvent {
        a: [0x11; 32],
        b: U256::from(1u64),
        c: keccak256("Hello World").into(),
        d: Vec::new(),
    };
    // topics are `(SELECTOR, a, keccak256(c))`
    assert_eq!(
        event.encode_topics_array::<3>(),
        [
            WordToken(MyEvent::SIGNATURE_HASH),
            WordToken(B256::repeat_byte(0x11)),
            WordToken(keccak256("Hello World"))
        ]
    );
    // dynamic data is `abi.encode(b, d)`
    assert_eq!(
        event.encode_data(),
        hex!(
            // b
            "0000000000000000000000000000000000000000000000000000000000000001"
            // d offset
            "0000000000000000000000000000000000000000000000000000000000000040"
            // d length
            "0000000000000000000000000000000000000000000000000000000000000000"
        ),
    );

    assert_event_signature::<LogNote>("LogNote(bytes4,address,bytes32,bytes32,uint256,bytes)");
    assert!(LogNote::ANONYMOUS);

    assert_event_signature::<MyEvent2>("MyEvent2((bytes))");
    assert!(!MyEvent2::ANONYMOUS);
}

fn assert_event_signature<T: SolEvent>(expected: &str) {
    assert_eq!(T::SIGNATURE, expected);
    assert_eq!(T::SIGNATURE_HASH, keccak256(expected));
}

Contracts/interfaces

use alloy_primitives::{Address, U256};
use alloy_sol_types::{sol, SolCall};
use hex_literal::hex;

sol! {
    /// Interface of the ERC20 standard as defined in [the EIP].
    ///
    /// [the EIP]: https://eips.ethereum.org/EIPS/eip-20
    #[derive(Debug, PartialEq)]
    interface IERC20 {
        event Transfer(address indexed from, address indexed to, uint256 value);
        event Approval(address indexed owner, address indexed spender, uint256 value);

        function totalSupply() external view returns (uint256);
        function balanceOf(address account) external view returns (uint256);
        function transfer(address to, uint256 amount) external returns (bool);
        function allowance(address owner, address spender) external view returns (uint256);
        function approve(address spender, uint256 amount) external returns (bool);
        function transferFrom(address from, address to, uint256 amount) external returns (bool);
    }
}

#[test]
fn contracts() {
    // random mainnet ERC20 transfer
    // https://etherscan.io/tx/0x947332ff624b5092fb92e8f02cdbb8a50314e861a4b39c29a286b3b75432165e
    let data = hex!(
        "a9059cbb"
        "0000000000000000000000008bc47be1e3abbaba182069c89d08a61fa6c2b292"
        "0000000000000000000000000000000000000000000000000000000253c51700"
    );
    let expected = IERC20::transferCall {
        to: Address::from(hex!("8bc47be1e3abbaba182069c89d08a61fa6c2b292")),
        amount: U256::from(9995360000_u64),
    };

    assert_eq!(data[..4], IERC20::transferCall::SELECTOR);
    let decoded = IERC20::IERC20Calls::decode(&data, true).unwrap();
    assert_eq!(decoded, IERC20::IERC20Calls::transfer(expected));
    assert_eq!(decoded.encode(), data);
}