alloy_sol_types
Trait SolEventInterface

pub trait SolEventInterface: Sized {
    const NAME: &'static str;
    const COUNT: usize;

    // Required method
    fn decode_raw_log(
        topics: &[Word],
        data: &[u8],
        validate: bool,
    ) -> Result<Self>;

    // Provided method
    fn decode_log(log: &Log, validate: bool) -> Result<Log<Self>> { ... }
}

A collection of SolEvents.

Implementer’s Guide

It should not be necessary to implement this trait manually. Instead, use the sol! procedural macro to parse Solidity syntax into types that implement this trait.

Required Associated Constants

const NAME: &'static str

The name of this type.

const COUNT: usize

The number of variants.

Required Methods

fn decode_raw_log(topics: &[Word], data: &[u8], validate: bool) -> Result<Self>

Decode the events from the given log info.

Provided Methods

fn decode_log(log: &Log, validate: bool) -> Result<Log<Self>>

Decode the events from the given log object.

Object Safety

This trait is not object safe.