Trait alloy_sol_types::SolStruct

pub trait SolStruct: 'static {
    type Tuple<'a>: SolType<TokenType<'a> = Self::Token<'a>>;
    type Token<'a>: TokenSeq<'a>;

    const NAME: &'static str;
    const FIELDS: &'static [(&'static str, &'static str)];

    // Required methods
    fn to_rust<'a>(&self) -> <Self::Tuple<'a> as SolType>::RustType;
    fn new(tuple: <Self::Tuple<'_> as SolType>::RustType) -> Self;
    fn tokenize(&self) -> Self::Token<'_>;
    fn eip712_encode_data(&self) -> Vec<u8> ;

    // Provided methods
    fn encoded_size(&self) -> usize { ... }
    fn eip712_encode_type() -> Cow<'static, str> { ... }
    fn eip712_type_hash(&self) -> B256 { ... }
    fn eip712_hash_struct(&self) -> B256 { ... }
    fn eip712_signing_hash(&self, domain: &Eip712Domain) -> B256 { ... }
}

A Solidity Struct.

This trait is used to implement ABI and EIP-712 encoding and decoding.

Implementer’s Guide

We do not recommend implementing this trait directly. Instead, we recommend using the sol proc macro to parse a Solidity struct definition.

Note

Special attention should be paid to eip712_encode_type for complex Solidity types. Nested Solidity structs must properly encode their type.

To be clear, a struct with a nested struct must encode the nested struct’s type as well.

See EIP-712#definition-of-encodetype for more details.

Required Associated Types

type Tuple<'a>: SolType<TokenType<'a> = Self::Token<'a>>

The corresponding Tuple type, used for encoding/decoding.

type Token<'a>: TokenSeq<'a>

The corresponding Token type.

Required Associated Constants

const NAME: &'static str

The struct name.

Used in eip712_encode_type.

const FIELDS: &'static [(&'static str, &'static str)]

The field types and names. Type is a Solidity string, and must conform to the name of the Solidty type at the same index in the associated tuple.

Used in eip712_encode_type.

Required Methods

fn to_rust<'a>(&self) -> <Self::Tuple<'a> as SolType>::RustType

Convert to the tuple type used for ABI encoding and decoding.

fn new(tuple: <Self::Tuple<'_> as SolType>::RustType) -> Self

Convert from the tuple type used for ABI encoding and decoding.

fn tokenize(&self) -> Self::Token<'_>

Convert to the token type used for EIP-712 encoding and decoding.

fn eip712_encode_data(&self) -> Vec<u8>

EIP-712 encodeData https://eips.ethereum.org/EIPS/eip-712#definition-of-encodedata

Provided Methods

fn encoded_size(&self) -> usize

The size of the struct when encoded, in bytes

fn eip712_encode_type() -> Cow<'static, str>

EIP-712 encodeType https://eips.ethereum.org/EIPS/eip-712#definition-of-encodetype

fn eip712_type_hash(&self) -> B256

EIP-712 typeHash https://eips.ethereum.org/EIPS/eip-712#rationale-for-typehash

fn eip712_hash_struct(&self) -> B256

EIP-712 hashStruct https://eips.ethereum.org/EIPS/eip-712#definition-of-hashstruct

fn eip712_signing_hash(&self, domain: &Eip712Domain) -> B256

EIP-712 signTypedData https://eips.ethereum.org/EIPS/eip-712#specification-of-the-eth_signtypeddata-json-rpc