alloy_sol_types

Module utils

Utilities used by different modules.

Functions

next_multiple_of_32

See usize::next_multiple_of.

words_for

Calculates the padded length of a slice by rounding its length to the next word.

words_for_len

Calculates the padded length of a slice of a specific length by rounding its length to the next word.