Trait alloy_sol_types::SolType

pub trait SolType {
    type RustType: Encodable<Self> + 'static;
    type TokenType<'a>: TokenType<'a>;

    const ENCODED_SIZE: Option<usize> = _;
    const DYNAMIC: bool = _;

    // Required methods
    fn sol_type_name() -> Cow<'static, str>;
    fn type_check(token: &Self::TokenType<'_>) -> Result<()>;
    fn detokenize(token: Self::TokenType<'_>) -> Self::RustType;
    fn eip712_data_word(rust: &Self::RustType) -> Word;
    fn encode_packed_to(rust: &Self::RustType, out: &mut Vec<u8>);

    // Provided methods
    fn encoded_size(_rust: &Self::RustType) -> usize { ... }
    fn tokenize<E>(rust: &E) -> Self::TokenType<'_>
       where E: Encodable<Self> { ... }
    fn eip712_encode_type() -> Option<Cow<'static, str>> { ... }
    fn encode_packed(rust: &Self::RustType) -> Vec<u8>  { ... }
    fn encode_single(rust: &Self::RustType) -> Vec<u8>  { ... }
    fn encode<'a>(rust: &'a Self::RustType) -> Vec<u8> 
       where Self::TokenType<'a>: TokenSeq<'a> { ... }
    fn encode_params<'a>(rust: &'a Self::RustType) -> Vec<u8> 
       where Self::TokenType<'a>: TokenSeq<'a> { ... }
    fn hex_encode<'a>(rust: &'a Self::RustType) -> String
       where Self::TokenType<'a>: TokenSeq<'a> { ... }
    fn hex_encode_single(rust: &Self::RustType) -> String { ... }
    fn hex_encode_params<'a>(rust: &'a Self::RustType) -> String
       where Self::TokenType<'a>: TokenSeq<'a> { ... }
    fn decode<'de>(data: &'de [u8], validate: bool) -> Result<Self::RustType>
       where Self::TokenType<'de>: TokenSeq<'de> { ... }
    fn decode_params(data: &[u8], validate: bool) -> Result<Self::RustType>
       where for<'de> Self::TokenType<'de>: TokenSeq<'de> { ... }
    fn decode_single(data: &[u8], validate: bool) -> Result<Self::RustType> { ... }
    fn hex_decode(data: &str, validate: bool) -> Result<Self::RustType>
       where for<'de> Self::TokenType<'de>: TokenSeq<'de> { ... }
    fn hex_decode_single(data: &str, validate: bool) -> Result<Self::RustType> { ... }
    fn hex_decode_params(data: &str, validate: bool) -> Result<Self::RustType>
       where for<'de> Self::TokenType<'de>: TokenSeq<'de> { ... }

}

A Solidity Type, for ABI encoding and decoding

This trait is implemented by types that contain ABI encoding and decoding info for Solidity types. Types may be combined to express arbitrarily complex Solidity types.

use alloy_sol_types::{SolType, sol_data::*};

type DynUint256Array = Array<Uint<256>>;
assert_eq!(&DynUint256Array::sol_type_name(), "uint256[]");

type Erc20FunctionArgs = (Address, Uint<256>);
assert_eq!(&Erc20FunctionArgs::sol_type_name(), "(address,uint256)");

type LargeComplexType = (FixedArray<Array<Bool>, 2>, (FixedBytes<13>, String));
assert_eq!(&LargeComplexType::sol_type_name(), "(bool[][2],(bytes13,string))");

These types are zero cost representations of Solidity types. They do not exist at runtime. They ONLY contain information about the type, they do not carry any data.

Implementer’s Guide

We do not recommend implementing this trait directly. Instead, we recommend using the crate::sol proc macro to parse a Solidity structdef into a native Rust struct.

alloy_sol_types::sol! {
    struct MyStruct {
        bool a;
        bytes2 b;
    }
}

// This is the native rust representation of a Solidity type!
// How cool is that!
const MY_STRUCT: MyStruct = MyStruct { a: true, b: [0x01, 0x02] };

Required Associated Types

type RustType: Encodable<Self> + 'static

The corresponding Rust type.

type TokenType<'a>: TokenType<'a>

The corresponding ABI token type.

See implementers of TokenType.

Provided Associated Constants

const ENCODED_SIZE: Option<usize> = _

The encoded size of the type, if known at compile time

const DYNAMIC: bool = _

Whether the encoded size is dynamic.

Required Methods

fn sol_type_name() -> Cow<'static, str>

The name of the type in Solidity.

fn type_check(token: &Self::TokenType<'_>) -> Result<()>

Check a token to see if it can be detokenized with this type.

fn detokenize(token: Self::TokenType<'_>) -> Self::RustType

Detokenize.

fn eip712_data_word(rust: &Self::RustType) -> Word

Encode this data according to EIP-712 encodeData rules, and hash it if necessary.

Implementer’s note: All single-word types are encoded as their word. All multi-word types are encoded as the hash the concatenated data words for each element

https://eips.ethereum.org/EIPS/eip-712#definition-of-encodedata

fn encode_packed_to(rust: &Self::RustType, out: &mut Vec<u8>)

Non-standard Packed Mode ABI encoding.

See encode_packed for more details.

Provided Methods

fn encoded_size(_rust: &Self::RustType) -> usize

Calculate the encoded size of the data, counting both head and tail words. For a single-word type this will always be 32.

fn tokenize<E>(rust: &E) -> Self::TokenType<'_>where E: Encodable<Self>,

Tokenize

fn eip712_encode_type() -> Option<Cow<'static, str>>

The encoded struct type (as EIP-712), if any. None for non-structs.

fn encode_packed(rust: &Self::RustType) -> Vec<u8>

Non-standard Packed Mode ABI encoding.

This is different from normal ABI encoding:

types shorter than 32 bytes are concatenated directly, without padding or sign extension;

dynamic types are encoded in-place and without the length;

array elements are padded, but still encoded in-place.

More information can be found in the Solidity docs.

fn encode_single(rust: &Self::RustType) -> Vec<u8>

Encode a single ABI token by wrapping it in a 1-length sequence.

fn encode<'a>(rust: &'a Self::RustType) -> Vec<u8> where Self::TokenType<'a>: TokenSeq<'a>,

Encode an ABI sequence.

fn encode_params<'a>(rust: &'a Self::RustType) -> Vec<u8> where Self::TokenType<'a>: TokenSeq<'a>,

Encode an ABI sequence suitable for function parameters.

fn hex_encode<'a>(rust: &'a Self::RustType) -> Stringwhere Self::TokenType<'a>: TokenSeq<'a>,

Hex output of encode.

fn hex_encode_single(rust: &Self::RustType) -> String

Hex output of encode_single.

fn hex_encode_params<'a>(rust: &'a Self::RustType) -> Stringwhere Self::TokenType<'a>: TokenSeq<'a>,

Hex output of encode_params.

fn decode<'de>(data: &'de [u8], validate: bool) -> Result<Self::RustType>where Self::TokenType<'de>: TokenSeq<'de>,

Decode a Rust type from an ABI blob.

fn decode_params(data: &[u8], validate: bool) -> Result<Self::RustType>where for<'de> Self::TokenType<'de>: TokenSeq<'de>,

Decode a Rust type from an ABI blob.

fn decode_single(data: &[u8], validate: bool) -> Result<Self::RustType>

Decode a Rust type from an ABI blob.

fn hex_decode(data: &str, validate: bool) -> Result<Self::RustType>where for<'de> Self::TokenType<'de>: TokenSeq<'de>,

Decode a Rust type from a hex-encoded ABI blob.

fn hex_decode_single(data: &str, validate: bool) -> Result<Self::RustType>

Decode a Rust type from a hex-encoded ABI blob.

fn hex_decode_params(data: &str, validate: bool) -> Result<Self::RustType>where for<'de> Self::TokenType<'de>: TokenSeq<'de>,

Decode a Rust type from a hex-encoded ABI blob.